apps/second-brain-offline/pyproject.toml
[project]
name = "second-brain-offline-course"
version = "0.1.0"
description = "Self-paced course on production LLMs and RAG by teaching you how to build an AI assistant on top of your Notion second brain."
readme = "README.md"
authors = [
    {name = "Paul Iusztin", email = "p.b.iusztin@gmail.com"},
]
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.3",
    "datasets>=2.14.4",
    "loguru>=0.7.3",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.7.0",
    "pymongo>=4.4.0",
    "zenml[server]>=0.73.0",
    "boto3>=1.36.0",
    "langchain>=0.3.14",
    "langchain-mongodb>=0.4.0",
    "langchain-openai>=0.3.0",
    "ipykernel>=6.29.5",
    "crawl4ai>=0.3.745",
    "langchain-huggingface>=0.1.2",
    "matplotlib>=3.10.0",
    "litellm>=1.59.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.7.2",
]

[tool.pip]
extra-index-url = "https://download.pytorch.org/whl/cpu/torch_stable.html"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/second_brain_offline"]

[tool.ruff]
target-version = "py312"
